# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SPK API",
        "description": "Документация API для системы учёта СПК Земледелец",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/warehouses": {
            "get": {
                "tags": [
                    "WarehousesPost"
                ],
                "summary": "Получение списка записей складов",
                "operationId": "a512782cc50f634d2b97f238afc69ecf",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "some title"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "enum": [
                                                            "Зерновой",
                                                            "Склад запчастей",
                                                            "Общий"
                                                        ],
                                                        "example": "Зерновой"
                                                    },
                                                    "status": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "area": {
                                                        "type": "integer",
                                                        "example": 2000
                                                    },
                                                    "max_historical_load": {
                                                        "type": "integer",
                                                        "example": 2000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "WarehousesPost"
                ],
                "summary": "Создание записи о складе",
                "operationId": "e525d442d2056d52bb883b569e097d9c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "some title"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "Зерновой",
                                                    "Склад запчастей",
                                                    "Общий"
                                                ],
                                                "example": "Зерновой"
                                            },
                                            "area": {
                                                "type": "integer",
                                                "example": 2000
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "max_historical_load": {
                                                "type": "integer",
                                                "example": 2000
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "some title"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "enum": [
                                                        "Зерновой",
                                                        "Склад запчастей",
                                                        "Общий"
                                                    ],
                                                    "example": "Зерновой"
                                                },
                                                "area": {
                                                    "type": "integer",
                                                    "example": 2000
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "max_historical_load": {
                                                    "type": "integer",
                                                    "example": 2000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/warehouses/{warehouse}/grains": {
            "get": {
                "tags": [
                    "WarehousesPost"
                ],
                "summary": "Получение остатков по зерновым культурам на складе",
                "operationId": "d24069d5558894b783acc25487d3e9d0",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "description": "ID склада",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "warehouse_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "grain_type_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "amount": {
                                                "type": "integer",
                                                "example": 1500
                                            },
                                            "grain_type": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Пшеница"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Склад не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/warehouses/{warehouse}/shipments": {
            "get": {
                "tags": [
                    "GrainShipments"
                ],
                "summary": "Получение истории отгрузок по складу",
                "operationId": "ddca4dbc0e05eb4ccabc09a7d962cc3e",
                "parameters": [
                    {
                        "name": "warehouse",
                        "in": "path",
                        "description": "ID склада",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список отгрузок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "warehouse_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "grain_type_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "grain_type": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Пшеница"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "volume": {
                                                "type": "number",
                                                "example": 1000
                                            },
                                            "shipment_date": {
                                                "type": "string",
                                                "example": "2025-06-15"
                                            },
                                            "driver": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Пётр Петров"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grain-shipments": {
            "post": {
                "tags": [
                    "GrainShipments"
                ],
                "summary": "Добавление отгрузки зерна",
                "operationId": "999929b57e0b51ec18f0e55d1879f7e5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "warehouse_id",
                                    "grain_type_id",
                                    "volume",
                                    "shipment_date"
                                ],
                                "properties": {
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grain_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "volume": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1000
                                    },
                                    "shipment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-15"
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Отгрузка добавлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отгрузка добавлена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/grain-shipments/{shipment}": {
            "put": {
                "tags": [
                    "GrainShipments"
                ],
                "summary": "Обновление отгрузки зерна",
                "operationId": "d3f9540903a74dc3174dea47d1efa8b5",
                "parameters": [
                    {
                        "name": "shipment",
                        "in": "path",
                        "description": "ID отгрузки",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "warehouse_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grain_type_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "volume": {
                                        "type": "number",
                                        "example": 950
                                    },
                                    "shipment_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-17"
                                    },
                                    "vehicle_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "driver_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Отгрузка обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отгрузка обновлена"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "volume": {
                                                    "type": "number",
                                                    "example": 950
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-06-01T12:00:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Получение списка ролей",
                "operationId": "a0daa85aa159b805a35a61e6a6e2be29",
                "responses": {
                    "200": {
                        "description": "Список ролей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Водитель"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/license-categories": {
            "get": {
                "tags": [
                    "LicenseCategories"
                ],
                "summary": "Получение списка категорий водительских прав",
                "operationId": "d8dbdbca38e0ca03434011bce0a47872",
                "responses": {
                    "200": {
                        "description": "Список категорий",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "code": {
                                                "type": "string",
                                                "example": "B"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}